Modify DatabaseDiary and add "test" file
Added the edit and remove methods to DD.

Updated DD's query method to return an array of objects.
The objects have two properties, key and value. Sadly we
have to use cursors for this. I do not like cursors that much?

Added my crappy test file to the repo. This is like the worst
"testing" file ever. I should probably just find a JavaScript
testing solution.